+++ libr/asm/meson.build
@@ -53,7 +53,6 @@ r_asm_sources = [
   join_paths('p','asm_v810.c'),
   join_paths('p','asm_v850.c'),
   join_paths('p','asm_v850_gnu.c'),
-  join_paths('p','asm_vax.c'),
   join_paths('p','asm_x86_as.c'),
   join_paths('p','asm_x86_cs.c'),
   join_paths('p','asm_x86_nasm.c'),
@@ -141,7 +140,6 @@ r_asm_sources = [
   join_paths('arch','tms320','tms320_dasm.c'),
   join_paths('arch','v810','v810_disas.c'),
   join_paths('arch','v850','v850_disas.c'),
-  join_paths('arch','vax','vax-dis.c'),
   # join_paths('arch','whitespace','wsdis.c'),
   # join_paths('arch','xap','dis.c'),
   #join_paths('arch','z80','expressions.c'),
